SIMOV-API/requirements.txt
##INSTALL DEPENDENCIAS PARA
#MYSQL DRIVER
#SQLSERVER DRIVER
#FLASK
#DOTENV
#REQUEST
#REDIS (SALVANDO HISTORICO EM REDIS)
#SWAGGER

flask==2.2.3
python-dotenv==1.0.0
requests==2.31.0
threaded==4.1.0
urllib3==1.26.6
status==0.2.5
flask-redis==0.4.0 
redis==4.5.5
Flask-Caching==2.0.2
flask_swagger_ui==4.11.1
mysql-connector==2.2.9 
retry==0.9.2
pyodbc==5.1.0


########################
########################
################# REQUERIMENTS